xresources: DPI change, lots of font tweaking & urxvt setup
This is a confusing commit. The most important change is the DPI
change from 128 to 96 to get better font rendering after I started
using a different font (fantasque sans mono).

I haven't been using urxvt that much, so the configuration got a
little bit out of control. And there's a whole lot of misc changes
about it here.